requirements.txt
boto3
coverage
python-dateutil